\name{LeerFotos-package}
\alias{LeerFotos-package}
\alias{LeerFotos}
\docType{package}
\title{
Leer Metadatos de Fotos *.JPG
}
\description{
se leen las fotografias y se guardan los datos en un archivo .csv en la carpeta fuente.
}
\details{
\tabular{ll}{
Package: \tab LeerFotos\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-13\cr
License: \tab GPL(>=2.0)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Camilo A Herrera Rozo 
Maintainer: Camilo A Herrera Rozo  to <hr.camilo@gmail.com>

}
\references{

}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
